Make join behave idempotently
When the joining cluster's secret already exists in the host cluster, we check the existing kubefed cluster and compare the control plane urls of the joining cluster and the existing kubefed cluster. If they are the same we re-use the same secret (possibly should refresh it). If they are different we exit with an error.